# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RescueBase API",
    "description": "API for Sanitätslager, Rucksackchecks, Chargen and Nachfüllaufträge.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Reverse-proxied API base path"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rb_session"
      }
    },
    "schemas": {
...
      "MpgRecord": {
        "type": "object",
        "additionalProperties": true
      },
...
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/mpg/documents": {
      "get": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgDocumentsController_list",
        "parameters": [
          {
            "name": "modelId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deviceId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "personId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trainingId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "incidentId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document versions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MpgRecord"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgDocumentsController_upload",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "modelId": {
                    "type": "string"
                  },
                  "deviceId": {
                    "type": "string"
                  },
                  "personId": {
                    "type": "string"
                  },
                  "trainingId": {
                    "type": "string"
                  },
                  "incidentId": {
                    "type": "string"
                  },
                  "previousVersionId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/documents/{id}": {
      "get": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgDocumentsController_download",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Binary report",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
    "/mpg/exports/devices/{id}.pdf": {
      "get": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgReportsController_device",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Binary report",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/exports/devices/{id}.zip": {
      "get": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgReportsController_archive",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Binary report",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/exports/devices/{id}/label.pdf": {
      "get": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgReportsController_label",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Binary report",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/exports/trainings/{id}.pdf": {
      "get": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgReportsController_training",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Binary report",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/exports/people/{id}.pdf": {
      "get": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgReportsController_person",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Binary report",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  }
}